Add files via upload
Module "instructions" contains long texts. Module "ruffier" contains the calculation of the result. The module "seconds" contains a class for timer counting. And the file "main_app" contains the main code. The program has 5 screens which are the instructions and fields for inputting information that is needed. In general, program calculates the user's heart beating and shows the result in the end and a possible advice